==> 07_assess_data.Rmd <==
---
title: "05_data_assessment"
author: "Morten Gade"
date: "11/29/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# ------------------------------------------------- RATINGS -------------------------------------------------
```{r cars}
# packages
pacman::p_load(tidyverse, dplyr, stringdist)
#library(tidyverse, dplyr, stringdist)

# load data
data <- read_csv("./data/folketinget_1990_2021_raw.csv")

# sample ~200 values across different years
set.seed(124)
samples <- data %>% group_by(Year) %>% sample_n(round(200/length(unique(data[["Year"]])))) %>% ungroup() %>% select(id, text, doc_id, Year, Name)

# assign rater
samples$who <- ifelse(samples$doc_id %in% ((sample(samples$doc_id, nrow(samples)/2))), "Morten", "Gustav")

```

```{r}
################################################### FUNCTION ######################################################  

assessment <- function(df){
  # list of ids
  doc_ids <- df$doc_id %>% sort()
  # Empty data frame
  ratings <- data.frame(doc_id = as.character(), split = as.character(), assignment = as.character())
  
  # Loop through ids
  i = 1
  for(doc_id in doc_ids){
    cat(paste0("Document number: ", i," out of ", length(doc_ids), "\n"))
    cat(paste0("PDF name:\n", df$id[df$doc_id == doc_id],".pdf", "\n"))
    cat(paste0("Assigned to:\n ", df$Name[df$doc_id == doc_id], "\n"))
    cat(paste0("Text: \n \n ---------------------------------------------------------------- \n ", 
               df$text[df$doc_id == doc_id]), "\n ----------------------------------------------------------------\n")
    split = readline(prompt = paste0("Is the text in ", doc_id, " split correctly? "))
    assignment = readline(prompt = paste0("Is the text in ", doc_id, " assigned correctly? "))
    row <- data.frame(doc_id = doc_id, split = split, assignment = assignment)
    ratings <- rbind(ratings, row)
    i = i+1
  }
  return(ratings)
}

```

```{r}
############################################### MORTEN RATING ######################################################
# morten <- samples %>% filter(who == "Morten")
# write.csv(morten, "./data/samples_morten.csv")

# 1 = Yes
# 2 = No

morten <- read_csv("./data/samples_morten.csv")
morten_ratings <- assessment(df = morten)

backup <- morten_ratings

morten_ratings %>% group_by(split) %>% summarise(n())
backup %>% group_by(split) %>% summarise(n())
backup %>% group_by(assignment, split) %>% summarise(n())

backup$split[backup$doc_id %in% c(20549, 40888, 64441, 67045, 106517, 133896, 164411, 197794, 211280, 219243, 353798, 391524, 456369, 532239, 798246, 805146)] <- 1

backup %>% filter()

write_csv(backup, "./data/morten_ratings.csv")

```

```{r}
############################################### GUSTAV RATING ######################################################
# gustav <- samples %>% filter(who == "Gustav")
# write.csv(gustav, "./data/samples_gustav.csv")

# 1 = Yes
# 2 = No

gustav <- read_csv("./data/samples_gustav.csv")
gustav_ratings <- assessment(df = gustav[81:96,])

# "Arbejdsministeren" mangler i titler
# Procedural phrases: "kort bemærkning"
# Bør også fjerne næstformand og tredtje formand og fjerde formand 
# "kulturministeren" fangede den ikke

```

```{r}
# writing txt file with pdf samples listed

# lapply(morten$id, write, "morten_pdf.txt", append=TRUE)
# lapply(gustav$id, write, "gustav_pdf.txt", append=TRUE)
```

```{r}
############################################### RATING RESULTS ######################################################

agg_ratings <- rbind(read_csv("./data/gustav_ratings.csv"), read_csv("./data/morten_ratings.csv")) %>% 
  group_by(assignment, split) %>% 
  summarise(n = n()) %>% 
  mutate(assignment = ifelse(assignment == 1, 0, 1),
         split = ifelse(split == 1, 0, 1)) %>% 
  ungroup() %>% 
  add_row(assignment = 1, split = 0, n = 0)

# by year

ratings <- rbind(read_csv("./data/gustav_ratings.csv"), read_csv("./data/morten_ratings.csv")) %>% 
  left_join(samples, by = "doc_id") %>% 
  mutate(split = ifelse(split == 1, 0, 1),
         assignment = ifelse(assignment == 1, 0, 1),
         agg = ifelse(split == 0 & assignment == 0, 0, 1)) %>% 
  group_by(Year) %>% 
  summarise(wrong = sum(agg))

ggplot(data = ratings)+
  geom_col(aes(x=Year,y=wrong))+
  geom_smooth(aes(x=Year,y=wrong), method = "lm", se=F, )+
  theme_minimal()+
  labs(y = "No. of documents", x = "Year", title = "Documents either assigned or split incorrectly")

```



# --------------------------------------------------- PLOTTING -------------------------------------------------
```{r}

# Setting the theme
theme_set(theme_minimal())

# Function for setting color theme
Darjeeling1 <- function(no_of_colors){
  colors = wesanderson::wes_palette("Darjeeling1", no_of_colors, type = "continuous")
  return(colors)
}

```

```{r}
################################ No. of speeches per party ###################################

data %>% 
  count(Parti) %>%
  ggplot()+
  aes(x = reorder(Parti, -n), y = n, fill = Parti, label = n)+
  geom_col(width=0.5, position='dodge') +
  scale_fill_manual(values = Darjeeling1(length(unique(data$Parti))))+
  geom_text(position=position_dodge(0.5), vjust=-.50) + 
  labs(x="Party", title = "No. of speeches per party") +
  ylim(0, 110000)+
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1),
        legend.position = "none",
        axis.title.y = element_blank())

```

```{r}
################################## NUMBER OF SPEECHES PR. MEMBER PR. YEAR ##########################################
# Wrangling  
data %>% group_by(Year, Name, Parti) %>% count() %>% 
  ungroup() %>% group_by(Year) %>% 
  summarise(mean = mean(n),
            median = median(n),
            fifth = quantile(n, probs = 0.05),
            nintyfifth = quantile(n, probs = 0.95)
            ) %>% pivot_longer(!Year) %>% 
  
  # Plotting
  ggplot()+
  aes(x = Year, y = value, color = name, linetype = name)+
  geom_line()+
  guides(linetype = FALSE)+
  scale_color_manual(labels = c("5th %ile", "Mean", "Median", "95th %ile"),
                     values = wesanderson::wes_palette("Darjeeling1", n = 4))+
  labs(x = "Year", 
       y = "No. of Speeches", 
       title = "No. of speeches per person")+
  theme(legend.title = element_blank(), 
        legend.key = element_rect(color = alpha("black", 0.2)),
        axis.title.y = element_blank())

```

```{r}
############################# LENGTH OF SPEECHES #######################################
# Wrangling  
data %>% group_by(Year) %>% 
  summarise(mean = mean(n_words),
            median = median(n_words),
            fifth = quantile(n_words, probs = 0.05),
            nintyfifth = quantile(n_words, probs = 0.95)
            ) %>% pivot_longer(!Year) %>% 
  
  # Plotting
  ggplot()+
  aes(x = Year, y = value, color = name, linetype = name)+
  geom_line(size = 1)+
  guides(linetype = FALSE)+
  scale_color_manual(labels = c("5th %ile", "Mean", "Median", "95th %ile"),
                     values = Darjeeling1(4))+
  ylim(0,1500)+
  labs(x = "Year", 
       y = "Speech length", 
       title = "Speech length (no. of words)")+
  theme(legend.title = element_blank(), 
        legend.key = element_rect(color = alpha("black", 0.2)),
        axis.title.y = element_blank())
  
```

```{r}
############################ ACCUMULATED NO. OF CHARACTERS FOR EACH PARTY ###################################
data %>% group_by(Parti) %>% 
  summarize(sum = sum(n_words)) %>%  
  ggplot()+
  aes(x = reorder(Parti, -sum), y = sum, fill = Parti)+
  geom_col(width=0.5, position='dodge') +
  scale_fill_manual(values = Darjeeling1(16) )+
  labs(y="Accumulated characters in speeches", x="Parti", title = "Total length of speeches for each party") +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1),
        legend.position = "none")

```

```{r}
####################################### CHECKING WHETHER THERE ARE ANY SIMILAR NAMES #############################

# Creating stringdist matrix for restaurant names
distmatrix<-stringdist::stringdistmatrix(unique(data$Name),unique(data$Name), useNames=TRUE ,method = "osa")

# Converting to dataframe
distmatrixdf<- as.data.frame(distmatrix)

# Making a row with restaurant names
distmatrixdf$Restaurant_Name <- rownames(distmatrixdf)

# Filtering out all links with less than 5 in distance and eyeballing the results to decide what links to merge
dist <- distmatrixdf %>% 
  pivot_longer(cols = everything(vars = distmatrixdf$Restaurant_Name)) %>% 
  filter(value <= 5 & value != 0)

```
