==> 08_wordshoal.Rmd <==
---
title: "wordshoal"
author: "Gustav Helms"
date: "11/8/2021"
output: html_document
---

```{r}
# loading packages
pacman::p_load(tidyverse, corpus, devtools, lubridate)

# Installing the latest version of quanteda and wordshoal
devtools::install_github("https://github.com/quanteda/quanteda")
remotes::install_github("kbenoit/wordshoal")
devtools::install_github("quanteda/quanteda.corpora")

# Loading the packages
library(quanteda)
library(wordshoal)
library(quanteda.textmodels)
library(quanteda.corpora)

# stopwords
#stopwords_da <- stopwords::stopwords("da")

stopwords <- read.delim("https://gist.githubusercontent.com/berteltorp/0cf8a0c7afea7f25ed754f24cfc2467b/raw/305d8e3930cc419e909d49d4b489c9773f75b2d6/stopord.txt", sep = "\n", col.names = "stopword") %>% as.list() %>% .[[1]]

# Defining how many cores to use
Sys.setenv("OMP_NUM_THREADS" = 16)

```


```{r}
# Importing data
df <- read_csv("./data/folketinget_1990_2021_raw.csv") %>% arrange(Month)
df$Month <- as.character(df$Month)

#subset <- df %>% filter(Month %in% c("1991-10-01", "1990-11-01", "1990-12-01"))
```



# --------------------------------------------------------------------------------------------------------------
# Inspecting the data frame

```{r}
# Checking party labels
df %>% group_by(Parti) %>% count() %>% arrange(n) # sanity check

summary(df$Date)


```

# --------------------------------------------------------------------------------------------------------------
# Implementing the wordshoal function on our own data

```{r}
##############  TESTING NEW FUNCTION ############## 

get_theta <- function(data){
  cat(paste0("[ ] Shoaling Month ", data$Month[1], ". Month: ", i, " out of ", length(unique(df$Month)),"\n"))
  corpus = quanteda::corpus(data)
  # Converting to tokens and removing punctuation, numbers, and symbols
  toks <- tokens(corpus, remove_punct = TRUE, remove_numbers = TRUE, remove_symbol = TRUE) %>% 
  # Removing danish stopwords
  tokens_remove(pattern = c(stopwords))

  wordshoalfit <- wordshoal::textmodel_wordshoal(dfm, groups = docvars(corpus, "id"), 
                                                 authors = docvars(corpus, "Name"))
  author_positions <- summary(wordshoalfit)$estimated.author.positions
  author_positions$row_names <- rownames(author_positions)

  fitdf <- merge(author_positions,
               docvars(corpus), 
               by.x = "row_names", by.y = "Name")
  fitdf <- subset(fitdf, !duplicated(row_names))
  theta_df <<- rbind(theta_df, fitdf)
  i <<- i+1
}

testfunc <- function(data, year_list, party_list){
  
  # subsetting df based on arguments
  if(missing(data)){
    cat("No data argument")
  } else if(hasArg(year_list) & hasArg(party_list)){
    data = data %>% filter(Year %in% year_list & Parti %in% party_list)
  } else if(hasArg(party_list)){
    data = data %>% filter(Parti %in% party_list)
  } else if(hasArg(year_list)){
    data = data %>% filter(Year %in% year_list)
  }
  
  # creating empty df to append to
  theta_df <<- data.frame()
  
  # create nested list of dataframes to which the shoal-function is applied
  nested = data %>% split(.$Month) # change to Month variable
  lapply(names(nested), function(x) get_theta(nested[[x]]))
  filename = paste0("./data/wordshoal_estimates_", min(data$Year), "_", max(data$Year), ".csv")
  write_csv(theta_df, filename)
}


```

```{r}
# getting theta estimates for each speaker per month
i = 1
hep <- testfunc(data = df)

length(unique(df$Month))

```

# --------------------------------- PREVIOUS METHOD -------------------------------------
```{r}
############## corpus per month ###################
years = seq(2009,2021) # Years to loop through

df_subset <- df %>% filter(Year %in% years) 
months = unique(df_subset$month_year_id) # input list

i=1
wordshoaler <- function(month) {
  cat(paste0("[ ] Shoaling Month ", month, ". Month: ", i, " out of ", length(months), "\n"))
  shoal_df = df_subset %>% filter(month_year_id == paste0(month))
  corpuseret = quanteda::corpus(shoal_df)
  dfmmm <- dfm(tokens(corpuseret), remove_punct = TRUE, ignoredFeatures = stopwords("danish")) # matrix making
  wordshoalfit <- wordshoal::textmodel_wordshoal(dfmmm, groups = docvars(corpuseret, "id"), 
                                                 authors = docvars(corpuseret, "Name"))
  author_positions <- summary(wordshoalfit)$estimated.author.positions
  author_positions$row_names <- rownames(author_positions)

  fitdf <- merge(author_positions,
               docvars(corpuseret), 
               by.x = "row_names", by.y = "Name")
  fitdf <- subset(fitdf, !duplicated(row_names))
  thetascores <<- rbind(thetascores, fitdf)
  i <<- i + 1
}

thetascores <- data.frame()
purrr::map(months, wordshoaler)

# writing the estimates
filename = paste0("./data/wordshoal_results/wordshoal_estimates_", min(years), "_", max(years), ".csv")
write_csv(thetascores, filename)

```

```{r}
# Aggregating to get the positions of the mean positon of the parties
author_positions <- summary(wordshoalfit)$estimated.author.positions
author_positions$row_names <- rownames(author_positions)

fitdf <- merge(author_positions,
               docvars(corpus), 
               by.x = "row_names", by.y = "Name")
fitdf <- subset(fitdf, !duplicated(row_names))

```


# ------------------------ Creating overlap variable ----------------------------------------
```{r}
# Loading the results
pacman::p_load(overlapping)
#csv_files = list.files(pattern="*.csv", path = "data/wordshoal_results/", full.names = T) 
#df = map_df(csv_files, read_csv)
df_ws <- read_csv("./data/wordshoal_estimates_1990_2021.csv")
#df$time <- as.Date(paste(year(df$Date), month(df$Date), "01", sep = "-"))

```


```{r}
# FUNCTION
overlapper <- function(period){
  # Printing status
  print(paste0("[ ] Calculating overlaps for month: ", period, ". Number: ", i, " out of ", as.character(length(unique(df_ws$Month)))))
  
  # Subsetting the period and filtering out parties that only has 1 speech in a month
  subset <- df_ws %>% filter(Month == period) %>% group_by(Parti) %>% filter(n() > 1) %>% ungroup()
  
  # Only continue if there is more than one party with more than one speech
  if(length(unique(subset$Parti)) > 1){
  
    # Making a list of parties
    parties <- unique(subset$Parti)
    
    # Converting into wide format
    wide <- subset %>% pivot_wider(names_from = Parti, values_from = theta) 
    
    # Creating an empty list
    list <- list()
    # Adding the parties to the list
    for(party in parties){
      list <- append(list,na.omit(select(wide, party)))
    }
    
    # Calculating the overlaps
    overlaps <- overlap(list)
    
    # Wringling the data
    results <- data.frame(overlaps$OV) %>% rownames_to_column("Parties") %>% 
      separate(Parties, c("From","To"), "-")
    
    # Making an empty data frame to append to
    end_results = data_frame()
    
    # Wrangling the result data frame
    for(party in parties){
      hep <- results %>% filter(From == party | To == party) %>% mutate(
        copy_from = From,
        From = ifelse(From == party, From, To),
        To = ifelse(To == party, copy_from, To)
      ) %>% select(-copy_from)
      
      end_results <- rbind(end_results, hep)
    }
    
    # Adding month_year_id variable
    end_results$Month <- period
    
    i <<- i + 1
    return(end_results)
  }
}

```

```{r}
# Looping over the function
i = 1
heppe <- map_df(unique(df_ws$Month), overlapper) # change df name
write_csv(heppe, "./data/overlap_results_raw.csv")
```

```{r}
df_overlap <- read_csv("./data/overlap_results_raw.csv")
```

```{r}
# Adding the aggregated columns to the results and exporting the data
agg_ws <- df_ws %>% group_by(Month, Parti) %>% summarise(mean_theta = mean(theta), 
                                                           sd_theta = sd(theta))

joined <- left_join(df_overlap, agg_ws, by = c("From" = "Parti", "Month" = "Month")) %>% 
  rename(
    Parti = From,
    overlap_with = To,
    overlap_score = overlaps.OV
  ) %>% mutate(
    difference_score = 1-overlap_score
  ) %>% 
  select(Month, Parti, overlap_with, overlap_score, difference_score, mean_theta, sd_theta)


```

```{r}
# Preprocessing

######## Preprocessing #########3
# Agregating results into a single data frame - VERY MESSY i know
agg_overlap <- joined %>% group_by(Parti, Month) %>% 
  arrange(-overlap_score, .by_group = TRUE) %>% slice(1:3) %>% # rename df
  summarise(overlap_score = mean(overlap_score),
            difference_score = mean(difference_score)) 

agg <- left_join(x = agg_overlap, y = agg_ws, by = c("Parti" = "Parti", "Month" = "Month")) %>% group_by(Parti, Month) %>% slice(1) %>% ungroup() # Date might need to be changed to Month

# Removing NA and 0 values
agg <- agg %>% filter(mean_theta != 0 & sd_theta != 0) %>% filter(sd_theta < 4)

# Adding a lag column for the predictors - difference_score and sd_theta
agg <- agg  %>%  mutate(
  Inclusion = sd_theta,
  Differentiation = difference_score
) %>% group_by(Parti) %>% 
  mutate(
  Lag_inclusion = lag(Inclusion),
  Lag_differentiation = lag(Differentiation)
) 

agg <- na.omit(agg)

# Adding the coalition column 
hep <- df_ws %>% select(Month, Parti, coalition) %>% group_by(Month, Parti) %>% slice(1) %>% ungroup()

agg <- left_join(agg, hep, by = c("Parti" = "Parti", "Month" = "Month"))

```


```{r}
# Exporting 
write_csv(agg, "./data/agg_wordshoal_results_final.csv") # check output file name
```
