==> 09_statistical_modelling.Rmd <==
---
title: "baysian_modelling"
author: "Gustav Helms"
date: "11/17/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r}
# Loading packages
pacman::p_load(tidyverse, brms, here, patchwork, cmdstanr, wesanderson)

theme_set(theme_minimal())
```


```{r}
# load results
agg <- read_csv("./data/agg_wordshoal_results_final.csv") %>% 
  mutate(Inclusion = 1- (Inclusion/max(Inclusion)))
```


```{r}
# inspecting the outome variable
p1 <- ggplot(agg, aes(x = Inclusion)) + 
  geom_density()+
  geom_vline(xintercept = mean(agg$Inclusion), size=0.5, color="red", linetype = "dashed")+
  geom_text(aes(x=mean(agg$Inclusion)+0.05, label=paste0("Mean ",round(mean(Inclusion), digits = 2)), y = 5), size = 3)


p3 <- ggplot(agg, aes(x = Differentiation)) + 
  geom_density()+
  geom_vline(xintercept = mean(agg$Differentiation), size=0.5, color="red", linetype = "dashed")+
  geom_text(aes(x=mean(agg$Differentiation)+0.12, label=paste0("Mean ",round(mean(Differentiation), digits = 2)), y=3), size = 3)+
  labs(x = "Distinctiveness")

p4 <- ggplot(agg, aes(x = log(Differentiation))) + 
  geom_density()+
  geom_vline(xintercept = mean(log(agg$Differentiation)), size=0.5, color="red", linetype = "dashed")+
  geom_text(aes(x=mean(log(agg$Differentiation))+0.3, label=paste0("Mean ",round(mean(log(Differentiation)), digits = 2)), y=0.9), size = 3)+
  labs(x = "Log(Distinctiveness)")

(p1) / (p3 + p4) + plot_annotation(title = "Density plot of variables with and without transformation")
```

```{r, fig.width= 15, fig.height= 5}
#plotting the raw data
agg %>% 
  pivot_longer(cols = c("Inclusion", "Differentiation"), names_to = "Need") %>% 
  ggplot()+
  aes(x = Month, y = value, color = Need)+
  geom_line()+
  geom_point(size = 0.5)+
  geom_smooth()+
  scale_color_manual(values = wes_palette("Darjeeling1"))
  
  
```

```{r}
###################### DEFINING MODEL AND MAKING PRIOR-PREDICTIVE CHECKS ######################

f1 <- bf(Inclusion ~ 1 + Lag_inclusion + Lag_differentiation + (1|Parti), family = gaussian())

get_prior(f1, data = agg)

# Prior for intercept
mean(agg$Inclusion)
# mean of 0.7398839
sd(agg$Inclusion)
# sd of 0.09703646


# Finding prior for random intercept (with and withouth log transformation)
agg %>% group_by(Parti) %>%  # Withouth log
  summarise(mean = mean(Inclusion), 
            sd = sd(Inclusion)) %>% 
  summarise(mean_of_mean = mean(mean),mean_of_sd = mean(sd))
# mean = 0.7435319 and sd = 0.1082213

prior_m1 <- c(
  prior(normal(0.7398839, 0.2), class = Intercept),
  prior(normal(0, 0.2), class = b),
  prior(normal(0.7435319, 0.1082213), class = sd), # i still can't figure this out
  prior(normal(0.09703646, 0.2), class = sigma)
)

# Running the model
m1_prior <- brm(
  formula = f1,
  data = agg,
  family = gaussian(),
  prior = prior_m1, 
  sample_prior = "only",
  chains = 4,
  cores = 2,
  backend = "cmdstanr",
  file = here("models", "m1_prior")
  )

# Prior predictive check
pp_check(m1_prior, nsamples = 100)+
  labs(title = "M1: Prior Predictive Check")
```


```{r}
############################ RUNNING THE MODEL WITH DATA  #####################################
# Running the model
m1 <- brm(
  formula = f1,
  data = na.omit(agg),
  family = gaussian(),
  prior = prior_m1, 
  sample_prior = T,
  chains = 4,
  cores = 2,
  backend = "cmdstanr",
  file = here("models", "m1")
  )


# Posterior predictive check
pp_check(m1, nsamples = 100)+
  labs(title = "M1: Posterior Predictive Check")

```


```{r}
# Diagnostics 
# Summary output
print(m1)

#Checking the chains
plot(m1)
#Catterpillars look damn fine
```


```{r}
#---------------------------------prior-posterior update------------------------

#Prior posterior update checks
m1_samples <- posterior_samples(m1)
m1_samples %>% 
  select(c("b_Lag_inclusion", "b_Lag_differentiation", "b_Lag_inclusion:Lag_differentiation", "sd_Parti__Intercept", "sigma", "Intercept",
           "prior_Intercept", "prior_b", "prior_sd_Parti", "prior_sigma")) %>% 
  rename(
    b_lagInclusion = b_Lag_inclusion,
    b_lagDifferentiation = b_Lag_differentiation,
    b_Interaction = "b_Lag_inclusion:Lag_differentiation",
    b_sdParti = sd_Parti__Intercept,
    b_sigma = sigma,
    b_Intercept = Intercept, 
    prior_Intercept = prior_Intercept,
    prior_sdParti = prior_sd_Parti,
  ) %>% 
  mutate(
    prior_lagInclusion = prior_b,
    prior_lagDifferentiation = prior_b,
    prior_Interaction = prior_b
  ) %>% select(-prior_b) %>% 
  pivot_longer(cols = everything(),
               values_to = "value",
               names_to = "distribution"
               ) %>% 
  separate(distribution, c("Distribution", "Parameter"), sep = "_") %>% 
  mutate(Distribution = ifelse(Distribution == "b", "Posterior", "Prior")) %>% 
  ggplot(aes(x = value, fill = Distribution)) + 
  geom_density(alpha = 0.8) +
  facet_wrap(~Parameter, nrow = 4) +
  labs(title = "Prior-posterior updates")+
  theme_bw()+
  scale_fill_manual(values = wesanderson::wes_palette(name = "Darjeeling1")) + 
  theme(strip.text = element_text(size=15),
        axis.title=element_text(size=15),
        plot.title = element_text(size = 15, face = "bold", hjust = 0.5)) + 
  xlim(-1,1)
```

```{r}
#---------------------------------prior-posterior update - a little closer look ------------------------

#Prior posterior update checks
m1_samples <- posterior_samples(m1)
m1_samples %>% 
  select(c("b_Lag_inclusion", "b_Lag_differentiation",
            "prior_b")) %>% 
  rename(
    b_lagInclusion = b_Lag_inclusion,
    b_lagDifferentiation = b_Lag_differentiation
  ) %>% 
  mutate(
    prior_lagInclusion = prior_b,
    prior_lagDifferentiation = prior_b
  ) %>% select(-prior_b) %>% 
  pivot_longer(cols = everything(),
               values_to = "value",
               names_to = "distribution"
               ) %>% 
  separate(distribution, c("Distribution", "Parameter"), sep = "_") %>% 
  mutate(Distribution = ifelse(Distribution == "b", "Posterior", "Prior")) %>% 
  ggplot(aes(x = value, fill = Distribution)) + 
  geom_density(alpha = 0.8) +
  facet_wrap(~Parameter, nrow = 1) +
  labs(title = "M1: Prior-posterior updates")+
  theme_bw()+
  scale_fill_manual(values = wesanderson::wes_palette(name = "Darjeeling1")) + 
  theme(strip.text = element_text(size=15),
        axis.title=element_text(size=15),
        plot.title = element_text(size = 15, face = "bold", hjust = 0.5)) + 
  xlim(-1,1)
```

```{r}
# ------------------------------- Plotting conditional effects --------------------------------------------
cond_effect <- plot(conditional_effects(m1, spaghetti=F, 
                         method="fitted", nsamples=100), points = T)[[2]]

cond_effect+
  geom_line(color = wes_palette("Darjeeling1")[1], size = 1)+
  labs(title = "M1: Effect of lagged values of distinctiveness on inclusion", 
       x = "Lagged values of distinctiveness")

plot(conditional_effects(m1, spaghetti=F, 
                         method="predict", nsamples=100), points=T)
```



```{r}
#------------------------------------influential data points and LOO------------

#check for influential data-points
m2_no_meta <- add_criterion(m1, "loo", reloo = T)
#3 problematic observation(s) found.
loo <- (m2_no_meta$criteria[[1]])
hep <- data.frame(loo$pointwise)

plot(loo(m1))
#How to look at this plot?

```


```{r}
#-----------------------------------Results-------------------------------------

#checking the output of the model
print(m1)
mcmc_plot(m1) +
  theme(axis.text.y = element_text(hjust = 0))


#test your hypotheses
# Significant effect of lag inclusion?
hypothesis(m1, "Lag_inclusion < 0")
plot(hypothesis(m1, "Lag_inclusion < 0"))

# Significant effect of lag differentiation?
hypothesis(m1, "Lag_differentiation > 0")
plot(hypothesis(m1, "Lag_differentiation < 0"))

```

# ----------------------------- MODEL WITH DIFFERENTIATION -------------------------
```{r}
###################### DEFINING MODEL AND MAKING PRIOR-PREDICTIVE CHECKS ######################

f2 <- bf(Differentiation ~ 1 + Lag_differentiation +  Lag_inclusion + (1|Parti), family = lognormal())

get_prior(f2, data = agg)

# Prior for intercept and sigma
mean(log(agg$Differentiation))
# mean = -1.23593
sd(log(agg$Differentiation))
# sd = 0.3958038

# Finding prior for random intercept
means <- agg %>% group_by(Parti) %>% 
  summarise(mean = mean(sd_theta),
            sd = sd(sd_theta)) %>% 
  summarise(mean_of_mean = mean(mean),
            sd_of_mean = sd(mean))
# mean 0.569 and sd on 0.049


prior_m2 <- c(
  prior(normal(-1.23593, 0.2), class = Intercept),
  prior(normal(0, 0.2), class = b),
  prior(normal(0.569, 0.049), class = sd),
  prior(normal(0.3958038, 0.2), class = sigma)
)

# Running the model
m2_prior <- brm(
  formula = f2,
  data = agg,
  family = lognormal(),
  prior = prior_m2, 
  sample_prior = "only",
  chains = 4,
  cores = 2,
  backend = "cmdstanr",
  file = here("models", "m2_prior")
  )


# Prior predictive check
pp_check(m2_prior, nsamples = 100)+
  labs(title = "M2: Prior Predictive Check")
```


```{r}
############################ RUNNING THE MODEL WITH DATA  #####################################
# Running the model
m2 <- brm(
  formula = f2,
  data = agg,
  family = lognormal(),
  prior = prior_m2, 
  sample_prior = T,
  chains = 4,
  cores = 2,
  backend = "cmdstanr",
  file = here("models", "m2")
  )


# Posterior predictive check
pp_check(m2, nsamples = 100)+
  labs(title = "M2: Posterior Predictive Check")

```

```{r}
summary(m2)
```
```{r}
#---------------------------------prior-posterior update - a little closer look ------------------------

#Prior posterior update checks
m2_samples <- posterior_samples(m2)
m2_samples %>% 
  select(c("b_Lag_differentiation", "b_Lag_inclusion",
            "prior_b")) %>% 
  rename(
    b_lagInclusion = b_Lag_inclusion,
    b_lagDifferentiation = b_Lag_differentiation
  ) %>% 
  mutate(
    prior_lagInclusion = prior_b,
    prior_lagDifferentiation = prior_b
  ) %>% select(-prior_b) %>% 
  pivot_longer(cols = everything(),
               values_to = "value",
               names_to = "distribution"
               ) %>% 
  separate(distribution, c("Distribution", "Parameter"), sep = "_") %>% 
  mutate(Distribution = ifelse(Distribution == "b", "Posterior", "Prior")) %>% 
  ggplot(aes(x = value, fill = Distribution)) + 
  geom_density(alpha = 0.8) +
  facet_wrap(~Parameter, nrow = 1) +
  labs(title = "M2: Prior-posterior updates")+
  theme_bw()+
  scale_fill_manual(values = wesanderson::wes_palette(name = "Darjeeling1")) + 
  theme(strip.text = element_text(size=15),
        axis.title=element_text(size=15),
        plot.title = element_text(size = 15, face = "bold", hjust = 0.5)) + 
  xlim(-1,1)
```
```{r}
# ------------------------------- Plotting conditional effects --------------------------------------------
cond_effect_2 <- plot(conditional_effects(m2, spaghetti=F, 
                         method="fitted", nsamples=100), points = T)[[2]]

cond_effect_2+
  geom_line(color = wes_palette("Darjeeling1")[1], size = 1)+
  labs(title = "M2: Effect of lagged values of inclusion on distinctiveness", 
       x = "Lagged values of inclusion",
       y = "Distinctiveness")

plot(conditional_effects(m2, spaghetti=F, 
                         method="predict", nsamples=100), points=T)
```


```{r}
# Trace plot
plot(m2)+
  labs(title = "M2: Trace plot")
```


```{r}
#------------------------------------influential data points and LOO------------

#check for influential data-points
# m2_no_meta <- add_criterion(m2_no_meta, "loo", reloo = T)
#3 problematic observation(s) found.

plot(loo(m2))
#How to look at this plot?

```


```{r}
#-----------------------------------Results-------------------------------------

#checking the output of the model
print(m2)
mcmc_plot(m2) +
  theme(axis.text.y = element_text(hjust = 0))


#test your hypotheses
# Significant effect of lag inclusion?
hypothesis(m2, "Lag_inclusion < 0")
plot(hypothesis(m2, "Lag_inclusion < 0"))

# Significant effect of lag differentiation?
hypothesis(m2, "Lag_differentiation > 0")
plot(hypothesis(m2, "Lag_differentiation < 0"))

```
