==> 11_sgd_results.Rmd <==
---
title: "06_sgd_results"
author: "Morten Gade"
date: "11/21/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Markdown for sgd results, might be superfluous but at least it is more structured this way

# Inspecting the results
```{r}
#################### INSPECTING THE RESULTS ###########################
# Import the results
pacman::p_load(tidyverse)
ML_results <- read_csv("./data/cv_results_1990-2021_ny.csv")
ML_subset <- read_csv("./data/cv_results_subset.csv")
ML_coalition <- read_csv("./data/cv_results_1990-2021_coalition.csv")
ML_twoparty <- read_csv("./data/cv_results_1990-2021_twoparty.csv")

```

```{r}
# Plotting the results

ML_subset %>% 
  group_by(Year) %>% 
  summarise(kappa = mean(kappa_test)) %>% 
  ggplot()+
  aes(x = Year, y = kappa)+
  geom_point()+
  geom_line(color = wesanderson::wes_palette("Darjeeling1")[2])+
  geom_smooth(method = "lm", alpha = 0.1, color = wesanderson::wes_palette("Darjeeling1")[1], size = 0.5)+
  ylim(0, 1)+
  labs(y = "Cohen Kappa score", title = "Subset")+
  theme_minimal()

ML_results %>% 
  group_by(Year) %>% 
  summarise(kappa = mean(kappa_test)) %>% 
  ggplot()+
  aes(x = Year, y = kappa)+
  geom_point()+
  geom_line(color = wesanderson::wes_palette("Darjeeling1")[2])+
  geom_smooth(method = "lm", alpha = 0.1, color = wesanderson::wes_palette("Darjeeling1")[1], size = 0.5)+
  ylim(0, 1)+
  labs(y = "Cohen Kappa score", title = "Whole dataset")+
  theme_minimal()

ML_coalition %>% 
  group_by(Year) %>% 
  summarise(kappa = mean(kappa_test)) %>% 
  ggplot()+
  aes(x = Year, y = kappa)+
  geom_point()+
  geom_line(color = wesanderson::wes_palette("Darjeeling1")[2])+
  geom_smooth(method = "lm", alpha = 0.1, color = wesanderson::wes_palette("Darjeeling1")[1], size = 0.5)+
  ylim(0, 1)+
  labs(y = "Cohen Kappa score", title = "Coalition")+
  theme_minimal()

ML_twoparty %>% 
  group_by(Year) %>% 
  summarise(kappa = mean(kappa_test)) %>% 
  ggplot()+
  aes(x = Year, y = kappa)+
  geom_point()+
  geom_line(color = wesanderson::wes_palette("Darjeeling1")[2])+
  geom_smooth(alpha = 0.1, color = wesanderson::wes_palette("Darjeeling1")[1], size = 0.5)+
  ylim(0, 1)+
  labs(y = "Cohen Kappa score", title = "SD vs. Venstre")+
  theme_minimal()

```

```{r}

pacman::p_load(lme4)

coalition <- ML_coalition %>% group_by(Year) %>% summarise(coalition_kappa = mean(kappa_test), coalition_acc = mean(accuracy_score)) %>% pivot_longer(cols = c("coalition_kappa", "coalition_acc"), names_to = "metric", values_to = "score")
party <- ML_results %>% group_by(Year) %>% summarise(party_kappa = mean(kappa_test), party_acc = mean(accuracy_score)) %>% pivot_longer(cols = c("party_kappa", "party_acc"), names_to = "metric", values_to = "score")

sgd <- rbind(coalition, party) %>% mutate(class = ifelse(str_detect(metric, "coalition"), "coalition", "party"))

sgd %>% 
  ggplot()+
  aes(x = Year, y = score, color = metric)+
  geom_point()+
  geom_line(size = 0.5)+
  #geom_smooth(alpha = 0.1, size = 0.5)+
  facet_wrap(~class)+
  theme_minimal()

```

```{r}

hep <- read_csv("./data/folketinget_1990_2021_raw.csv")

hej <- hep %>% filter(Parti %in% c("Venstre", "Socialdemokratiet"))

per <- hej %>% group_by(Year, Parti) %>% summarize(n())

```

# ------------------------------------------ COMPARE WORDSHOAL AND ML --------------------------------------------------
```{r}

overlaps <- read_csv("~/bachelor_2021/data/overlap_results_raw.csv") %>%
  mutate(Year = lubridate::year(Month))

polarization <- overlaps %>% 
  group_by(Year) %>% 
  summarize(overlap = mean(overlaps.OV)) %>% 
  mutate(method = "Wordshoal",
         mean = 1 - overlap) %>% 
  select(Year, mean, method)

ML_results <- read_csv("./data/cv_results_1990-2021_ny.csv") %>% 
  group_by(Year) %>%
  summarise(mean = mean(kappa_test)) %>% 
  mutate(method = "Classifier") %>% 
  select(Year, mean, method)


comparison <- rbind(polarization, ML_results)

comparison %>% 
  ggplot()+
  aes(x = Year, y = mean, color = method)+
  geom_point()+
  geom_line()+
  scale_color_manual(values = wesanderson::wes_palette("Darjeeling1"))
  #ylim(0, 1)+
  labs(title = "Polarization")+
  theme_minimal()

cor.test(ML_results$mean, polarization$mean)
```
