==> 12_visualization.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r}
# loading packages
pacman::p_load(tidyverse,stringdist, overlapping, lubridate)
```

```{r}
# Importing data
df <- read_csv("data/folketinget_1990_2021_raw.csv") %>% filter(Year %in% c(2019, 2020))

# Color pallette
pal <- wesanderson::wes_palette("Darjeeling1", 14, type = "continuous")

# Setting up the theme
theme_set(theme_minimal())
```

# Creating figure for cor test
```{r}
# Loading cv results
cv <- read_csv("data/cv_results_1990-2021_ny.csv") %>% group_by(Year) %>% summarise(kappa = mean(kappa_test))

# Loading ws results
ws <- read_csv("data/agg_wordshoal_results_final.csv") 
ws <- ws %>% mutate(Year = lubridate::year(Month)) %>% group_by(Year) %>% summarise(ws = mean(Differentiation))

# plotting
left_join(cv,ws) %>% 
  filter(Year != c(1990,1991)) %>% 
  pivot_longer(cols = c("kappa","ws")) %>% 
  ggplot()+
  aes(x = Year, y = value, color = name)+
  geom_point()+
  geom_line()+
  ylim(0,1)+
  scale_color_manual(labels = c("Multivariate classifier", "Wordshoal"),
                     values = wesanderson::wes_palette("Darjeeling1"))+
  labs(title = "Polarization estimates of the two methods", color = "Methods")+
  theme(legend.position = "bottom")

cor.test(cv$kappa, ws$ws)
```





```{r}
# Mean position and sd of position per month and party
#agg <- ws_df %>% group_by(Parti, Month) %>% summarise(mean = mean(theta), sd = sd(theta)) %>% drop_na(sd)
agg <- ws_df %>% group_by(Parti, month_year_id) %>% summarise(mean = mean(theta), sd = sd(theta)) %>% drop_na(sd)

# Mean position of each member grouped by party
ws_df %>% group_by(row_names) %>% summarise(Parti = Parti, mean_theta = mean(theta)) %>% filter(!duplicated(mean_theta)) %>% filter(!row_names %in% c("Thor Pedersen", "Anders Fogh Rasmussen", "Søren Gade")) %>% 
  ggplot()+
  aes(x = mean_theta, y = Parti, color = Parti)+
  geom_point()+
  scale_color_manual(values = wesanderson::wes_palette("Darjeeling1", 10, type = "continuous"))+
  labs(title = "Wordshoal - Mean estimated position of each member divided by party", x = "Estimated position", y ="")+
  theme(legend.position = "none")
```

```{r}
# Aggregated pr year
agg %>%   
  mutate(Year = year(Month)) %>% group_by(Year, Parti) %>% summarise(mean = mean(mean)) %>% 
  ggplot()+
  aes(x = Year, y = mean, color = Parti)+
  geom_point()+
  geom_line()+
  scale_color_manual(values = wesanderson::wes_palette("Darjeeling1", 10, type = "continuous"))+
  labs(title = "Wordshoal - Estimated positions for parties per year", y = "Estimated position")
```



# ------------------------ Creating overlap variable ----------------------------------------
```{r}
# Loading the results
#csv_files = list.files(pattern="*.csv", path = "data/wordshoal_results/", full.names = T) 
#df = map_df(csv_files, read_csv)
df_ws <- read_csv("./data/wordshoal_results/wordshoal_estimates_2018_2021.csv")
#df$time <- as.Date(paste(year(df$Date), month(df$Date), "01", sep = "-"))

```


```{r}
# FUNCTION
overlapper <- function(period){
  # Printing status
  print(paste0("[ ] Calculating overlaps for month: ", period, ". Number: ", i, " out of ", as.character(length(unique(df$Month)))))
  
  # Subsetting the period and filtering out parties that only has 1 speech in a month
  subset <- df %>% filter(Month == period) %>% group_by(Parti) %>% filter(n() > 1) %>% ungroup()
  
  # Only continue if there is more than one party with more than one speech
  if(length(unique(subset$Parti)) > 1){
  
    # Making a list of parties
    parties <- unique(subset$Parti)
    
    # Converting into wide format
    wide <- subset %>% pivot_wider(names_from = Parti, values_from = theta) 
    
    # Creating an empty list
    list <- list()
    # Adding the parties to the list
    for(party in parties){
      list <- append(list,na.omit(select(wide, party)))
    }
    
    # Calculating the overlaps
    overlaps <- overlap(list)
    
    # Wringling the data
    results <- data.frame(overlaps$OV) %>% rownames_to_column("Parties") %>% 
      separate(Parties, c("From","To"), "-")
    
    # Making an empty data frame to append to
    end_results = data_frame()
    
    # Wrangling the result data frame
    for(party in parties){
      hep <- results %>% filter(From == party | To == party) %>% mutate(
        copy_from = From,
        From = ifelse(From == party, From, To),
        To = ifelse(To == party, copy_from, To)
      ) %>% select(-copy_from)
      
      end_results <- rbind(end_results, hep)
    }
    
    # Adding month_year_id variable
    end_results$Month <- period
    
    i <<- i + 1
    return(end_results)
  }
}

```

```{r}
# Looping over the function
i = 1
heppe <- map_df(unique(df_ws$Month), overlapper)

```

```{r}
# Looping over the function
i = 1
heppe <- map_df("2011_5", overlapper)
```

```{r}
# Adding the aggregated columns to the results and exporting the data
agg <- df %>% group_by(Month, Parti) %>% summarise(mean_theta = mean(theta), 
                                                           sd_theta = sd(theta))

joined <- left_join(heppe, agg, by = c("From" = "Parti", "Month" = "Month")) %>% 
  rename(
    Parti = From,
    overlap_with = To,
    overlap_score = overlaps.OV
  ) %>% mutate(
    difference_score = 1-overlap_score
  ) %>% 
  select(Month, Parti, overlap_with, overlap_score, difference_score, mean_theta, sd_theta)

write_csv(joined, "./data/agg_results_2009_2021.csv")
```


# ------------------------ Plotting results of overlap ----------------------------------------
```{r}
# Load results dataframe
results <- read_csv("./data/agg_results_2009_2021.csv")
```

```{r}
results %>%  group_by(Parti, Month) %>% summarise(mean = mean(difference_score)) %>% ungroup() %>% group_by(Month) %>% summarise(mean = mean(mean)) %>% ungroup() %>% 
  ggplot()+
  aes(x = Month, y = mean)+
  geom_point()+
  geom_line(color = wesanderson::wes_palette("Darjeeling1")[2])+
  geom_smooth(method = "lm", alpha = 0.1, color = wesanderson::wes_palette("Darjeeling1")[1], size = 0.5)+
  ylim(0, 1)+
  labs(x = "Time",y = "Mean difference score", title = "Wordshoal - Level of polarization in Folketinget")
  
```

```{r}
results %>%  group_by(Parti, Month) %>% summarise(mean = mean(difference_score)) %>% ungroup() %>% 
  ggplot()+
  aes(x = Month, y = mean)+
  geom_point()+
  geom_line(aes(color = Parti))+
  geom_smooth(method = "lm", alpha = 0.1, color = wesanderson::wes_palette("Darjeeling1")[1], size = 0.5)+
  ylim(0, 1)+
  scale_color_manual(values = wesanderson::wes_palette("Darjeeling1", 10, type = "continuous")[2:10])+
  facet_wrap(~Parti)+
  labs(x = "Time",y = "Mean difference score", title = "Wordshoal - Level of polarization in Folketinget per party")+
  theme(legend.position = "none")
  
```


# ------------------------------------- Plotting wordshoal results --------------------------------------------

```{r}
# Loading data
files <- list.files("./data/wordshoal_results/", pattern = "*.csv", full.names = T)
results <- map_df(files, read_csv)


```

```{r}
results %>% group_by(month_year_id) %>% count() %>% arrange(-n)

results %>% filter(month_year_id == "2017_10") %>% count(Parti)
```

```{r}
# Remove from visualization
to_keep <- c("Dansk Folkeparti", "Socialdemokratiet")

results %>% filter(month_year_id == "2017_10") %>% mutate(y_value = 0) %>%
  ggplot()+
  geom_density(aes(x = theta, fill = Parti), alpha =0.4)+
  geom_point(aes(x = theta, y = y_value, color = Parti), size = 0.5)+
  scale_fill_manual(values = wesanderson::wes_palette("Darjeeling1", n = 11, type = "continuous"))+
  scale_color_manual(values = wesanderson::wes_palette("Darjeeling1", n = 11, type = "continuous"))+
  xlim(-1.5, 1.5)+
  facet_wrap(~Parti)


results %>% filter(month_year_id == "2015_8" & !Parti %in% remove) %>% mutate(y_value = 0) %>%
  ggplot()+
  geom_density(aes(x = theta, fill = Parti), alpha = 0.4)+
  geom_point(aes(x = theta, y = y_value, color = Parti))+
  scale_fill_manual(values = wesanderson::wes_palette("Darjeeling1", n = 5, type = "continuous"))+
  scale_color_manual(values = wesanderson::wes_palette("Darjeeling1", n = 5, type = "continuous"))+
  xlim(-1.5, 1.5)
# Ud med 

```



```{r}
results %>% filter(month_year_id == "2019_10") %>% mutate(y_value = 1) %>% group_by(Parti) %>% 
  ggplot()+
  aes(x = theta, color = Parti) +
  stat_function(fun = dnorm)+
  xlim(-5,5)+
  scale_color_manual(values = wesanderson::wes_palette("Darjeeling1", n = 10, type = "continuous"))+
  theme(axis.title.y = element_blank(),
        axis.text.y = element_blank(),
        legend.position = "none")

ggplot(data.frame(x = c(-4, 4)), aes(x)) + 
  mapply(function(mean, sd, col) {
    stat_function(fun = dnorm, aes(x = ), col = col)
  }, 
  # enter means, standard deviations and colors here
  mean = c(0, 1, .5), 
  sd = c(1, .5, 2), 
  col = c('red', 'blue', 'green')
)

normal_curve <- function(mean, sd, col){
  stat_function(fun = dnorm, args = list(mean = mean, sd = sd), col = col)
}

means <- results %>% filter(month_year_id == "2019_10") %>% mutate(y_value = 1) %>% group_by(Parti) %>% 
  summarise(mean = mean(theta), sd = sd(theta)) %>% ungroup() %>% 
  mutate(col = wesanderson::wes_palette("Darjeeling1", n = 10, type = "continuous"))

ggplot(results) + 
  mapply(function(mean, sd, col) {
    stat_function(fun = dnorm, args = list(mean = mean, sd = sd), col = col, fill = col)
  }, 
  # enter means, standard deviations and colors here
  mean = means$mean,
  sd = means$sd, 
  col = means$col 
)+
  xlim(-1.5,1.5)
```
